+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 * Added `decode_hash`, `regexp` and `original` decoders
+* Fixed incorrect `Decoding::Failure` value comparisons
 
 ## [0.1.0] - 2024-05-08